+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+##### Added
+* Added support for the context API added in react `16.3.0`. See `React.createContext(...)`.
+
 #### Changed
 * 💥 **\[core\]** Remove support for the `componentWillReceiveProps` lifecycle method as it is being deprecated in
   React 16.3 and removed in React 17.x.